fix(update): check runtime at the state preparation boundary
