+++ src/cli/update-cli/update-command-run.ts
@@ -1,3 +1,7 @@
+import {
+  formatUnsupportedNodeVersionMessage,
+  isSupportedOpenClawNodeVersion,
+} from "../../../node-version.mjs";
 import { assertConfigWriteAllowedInCurrentMode } from "../../config/config.js";
 import { disableCurrentOpenClawUpdateLaunchdJob } from "../../daemon/launchd.js";
 import { mergeGatewayServiceEnv } from "../../daemon/service-env-merge.js";
@@ -41,9 +45,11 @@ import {
 } from "../../infra/update-run-ledger.js";
 import { summarizeUpdateStepFailure, type UpdateRunStep } from "../../infra/update-run-record.js";
 import type { UpdateRunResult, UpdateStepProgress } from "../../infra/update-runner.js";
+import { defaultRuntime } from "../../runtime.js";
 import { resolveOpenClawStateSqlitePath } from "../../state/openclaw-state-db.paths.js";
 import { assertOpenClawStateWriteAllowedAtPath } from "../../state/openclaw-state-ownership.js";
 import { VERSION } from "../../version.js";
+import { exitCliAfterOutput } from "../one-shot-exit.js";
 import { parseUpdateTimeoutMs, resolveUpdateRoot, type UpdateCommandOptions } from "./shared.js";
 import { suppressDeprecations } from "./suppress-deprecations.js";
 import {
@@ -250,6 +256,23 @@ export function readDevUpdateTarget(): DevUpdateTarget | undefined {
 }
 
 export async function prepareUpdateCommand(opts: UpdateCommandOptions) {
+  // Refuse before preflight can inspect write ownership or admit a live run ledger.
+  if (!process.versions.bun && !isSupportedOpenClawNodeVersion(process.versions.node)) {
+    const error = formatUnsupportedNodeVersionMessage(process.versions.node);
+    if (opts.json) {
+      defaultRuntime.writeJson({
+        status: "error",
+        mode: "unknown",
+        reason: "node-runtime-preflight",
+        error,
+        steps: [],
+        durationMs: 0,
+      });
+    } else {
+      defaultRuntime.error(`node-runtime-preflight: ${error}`);
+    }
+    exitCliAfterOutput(defaultRuntime, 1);
+  }
   const startedAt = Date.now();
   suppressDeprecations();
   const postCoreUpdateResume = process.env[POST_CORE_UPDATE_ENV] === "1";
